requirements.txt
pandas>=1.3.0
numpy>=1.20.0
statsmodels>=0.12.0
linearmodels>=4.25
openpyxl>=3.0.0
python-docx>=0.8.11
scipy>=1.7.0
